update inidicator implement
